+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## master
 
-- Added `Enumerator.produce` w/ `size:` support (Ruby 4.0).
+- Add `Array#rfind`.
+
+- Add `Enumerator.produce` w/ `size:` support (Ruby 4.0).
 
 ## 1.1.2 (2025-06-24)